fix: 🐛 unexpected feedback in "vertices" edge tool
Currently the `vertices` EdgeTool does not allow inserting new points
between exising points, only at the end of the edge.

This is unnatural for end-users, because one would expect
the tool to create a point
wherever they click on the edge.

Fix by using `insertVertex` instead of `setVertexAt` such that
a new vertex is guaranteed to be added.

Reference issue antvis/X6#1679